requirements.txt
click==6.7
gitpython==2.1.5
pygithub==1.35
termcolor==1.1.0
